삭제 시 비밀번호 검증추가

+++ src/main/java/org/example/scheduleforuser/service/ScheduleService.java
@@ -56,10 +56,14 @@ public class ScheduleService {
     }
 
     @Transactional
-    public void deleteSchedule(Long id) {
+    public void deleteSchedule(Long id, String password) {
         Schedule schedule = scheduleRepository.findById(id).orElseThrow(
                 ()-> new NoSuchElementException("\"해당 ID의 일정이 존재하지 않습니다:" + id)
         );
+
+        if (!schedule.getPassword().equalsIgnoreCase(password)) {
+            throw new IllegalArgumentException("비밀번호가 다릅니다.");
+        }
 
         scheduleRepository.delete(schedule);
     }
